minor updates based on comments
