+++ test/pageobjects/BasePage.js
@@ -1,5 +1,8 @@
 import { assert } from 'chai'
 import { locale, verifyElementCopy } from '../utils/mochaw'
+import { browserName, isRemoteBrowser } from '../main'
+import remote from 'selenium-webdriver/remote'
+import path from 'path'
 
 class BasePage {
   constructor(driver, waitAndFind) {
@@ -88,6 +91,29 @@ class BasePage {
       'Test Failed: Cobrand logo should be displayed'
     )
     verifyElementCopy(this.poweredBy(), 'powered by')
+  }
+
+  async upload(filename) {
+    const input = this.$('.onfido-sdk-ui-CustomFileInput-input')
+    const pathToTestFiles = '../resources/'
+    //the below line...if safari AND Local, ignore it...otherwise file uploads will fail.
+    if (browserName === 'safari' && isRemoteBrowser === false) {
+      console.log(
+        'Not creating a remote File Detector as I am uploading locally'
+      )
+    } else {
+      // This will detect local file, ref: https://www.browserstack.com/automate/node#enhancements-uploads-downloads
+      this.driver.setFileDetector(new remote.FileDetector())
+    }
+    const filePath = path.join(__dirname, pathToTestFiles + filename)
+    let sendKeysToElement
+    try {
+      console.log(`File being uploaded is ${filePath}`)
+      sendKeysToElement = input.sendKeys(path.join(filePath))
+    } catch (err) {
+      console.log('Just logging that I have caught an exception on upload')
+    }
+    return sendKeysToElement
   }
 }
